refactor: move findDuplicates to hardhat-utils
